requirements.txt
# Development
pep8
pyflakes
coverage
python-coveralls
# Requirements
iso8601
python-dateutil
pytz
requests
six
